requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
google-generativeai>=0.3.2
bcrypt>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pymongo>=4.9.0,<5.0.0
motor>=3.3.0

# Ferramentas de qualidade de código
pylint>=2.17.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0

# Dependências de teste (opcionais)
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx>=0.24.0
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
pytest
httpx
coverage